requirements.txt
python-dotenv
pytest
dotenv
browserstack-sdk
appium-python-client
pytest-selenium
selenium